+++ apps/web-giddh/src/app/actions/login.action.ts
@@ -1176,32 +1176,6 @@ export class LoginActions {
     }
 
     public finalNavigate(route: any, parameter?: any): void {
-        let isQueryParams: boolean;
-        if (screen.width <= 767 || isCordova) {
-            this._router.navigate(["/pages/mobile-home"]);
-        } else {
-            if (route.includes('?')) {
-                parameter = parameter || {};
-                isQueryParams = true;
-                const splittedRoute = route.split('?');
-                route = splittedRoute[0];
-                const paramString = splittedRoute[1];
-                const params = paramString?.split('&');
-                params?.forEach(param => {
-                    const [key, value] = param.split('=');
-                    parameter[key] = value;
-                });
-            }
-            if (isQueryParams) {
-                this._router.navigate([route], { queryParams: parameter });
-            } else {
-                this._router.navigate([route], parameter);
-            }
-        }
-        if (isCordova || isElectron) {
-            setTimeout(() => {
-                window.location.reload();
-            }, 200);
-        }
+        this._generalService.finalNavigate(route, parameter);
     }
 }
