+++ apps/web-giddh/src/app/shared/primary-sidebar/primary-sidebar.component.ts
@@ -237,7 +237,7 @@ export class PrimarySidebarComponent implements OnInit, OnChanges, OnDestroy {
         this.smartCombinedList$ = this.store.pipe(select(appStore => appStore.general.smartCombinedList), takeUntil(this.destroyed$));
         this.updateIndexDbSuccess$ = this.store.pipe(select(appStore => appStore.general.updateIndexDbComplete), takeUntil(this.destroyed$))
         this.store.pipe(select(state => state.session.activeCompany), takeUntil(this.destroyed$)).subscribe(selectedCmp => {
-            if (selectedCmp?.uniqueName !== this.selectedCompanyDetails?.uniqueName) {
+            if (selectedCmp && selectedCmp?.uniqueName === this.generalService.companyUniqueName) {
                 this.selectedCompany = observableOf(selectedCmp);
                 this.selectedCompanyDetails = selectedCmp;
                 this.companyInitials = this.generalService.getInitialsFromString(selectedCmp.name);
@@ -253,7 +253,7 @@ export class PrimarySidebarComponent implements OnInit, OnChanges, OnDestroy {
                     this.activeCompanyForDb.name = selectedCmp.name;
                     this.activeCompanyForDb.uniqueName = selectedCmp.uniqueName;
                 }
-                if (this.activeCompanyForDb?.uniqueName) {
+                if (this.generalService.companyUniqueName) {
                     this.dbService.getAllItems(this.activeCompanyForDb.uniqueName, 'accounts').subscribe(accountList => {
                         if (accountList?.length) {
                             if (window.innerWidth > 1440 && window.innerHeight > 717) {
@@ -261,10 +261,10 @@ export class PrimarySidebarComponent implements OnInit, OnChanges, OnDestroy {
                             } else {
                                 this.accountItemsFromIndexDB = accountList.slice(0, 5);
                             }
-                            this.changeDetectorRef.detectChanges();
                         } else {
                             this.accountItemsFromIndexDB = DEFAULT_AC;
                         }
+                        this.changeDetectorRef.detectChanges();
                     });
                 }
             }
@@ -415,16 +415,9 @@ export class PrimarySidebarComponent implements OnInit, OnChanges, OnDestroy {
         this.setOrganizationDetails(OrganizationType.Branch, details);
         this.companyService.getStateDetails(this.generalService.companyUniqueName).pipe(take(1)).subscribe(response => {
             if (response && response.body) {
-                if (screen.width <= 767 || isCordova) {
-                    window.location.href = '/pages/mobile-home';
-                } else if (isElectron) {
-                    this.router.navigate([response.body.lastState]);
-                    setTimeout(() => {
-                        window.location.reload();
-                    }, 200);
-                } else {
-                    window.location.href = response.body.lastState;
-                }
+                this.router.navigateByUrl('/dummy', { skipLocationChange: true }).then(() => {
+                    this.generalService.finalNavigate(response.body.lastState);
+                });
             }
         });
     }
